requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pyjwt[crypto]==2.9.0
python-multipart==0.0.12